move icon-list to icons
