+++ src/components/10-atoms/button-link/story.js
@@ -9,7 +9,7 @@ import {
 } from '@storybook/addon-knobs';
 import { html, render } from 'lit-html';
 import './index';
-import { iconList } from '../../../utils/icon-list';
+import { iconList } from '../icon/icon-list';
 import Readme from './README.md';
 
 const variantOptions = {
@@ -61,8 +61,8 @@ storyButtonLink.add('Button Link', () => {
   const template = html`
     <div
       style="${variants.includes('inverted')
-        ? `background-color: ${invertedBgs[variants]}; padding: 10px;`
-        : ''}"
+      ? `background-color: ${invertedBgs[variants]}; padding: 10px;`
+      : ''}"
     >
       <axa-button-link
         variant="${variants}"
