+++ src/components/10-atoms/input-file/story.js
@@ -2,7 +2,7 @@
 import { storiesOf } from '@storybook/html';
 import { text, select, boolean, withKnobs } from '@storybook/addon-knobs';
 import { html, render } from 'lit-html';
-import { iconList } from '../../../utils/icon-list';
+import { iconList } from '../icon/icon-list';
 import './index';
 import Readme from './README.md';
