+++ src/components/20-molecules/file-upload/react/story.jsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/html';
 import { text, number, select, withKnobs } from '@storybook/addon-knobs';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { iconList } from '../../../../utils/icon-list';
+import { iconList } from '../../../10-atoms/icon/icon-list';
 import AXAFileUploadReact from './AXAFileUploadReact';
 import Readme from '../README.md';
